backend/pyproject.toml
[project]
name = "guardianeye"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.41.0",
]
